refactor(VCarousel): add typescript tests, remove deprecated tests
